+++ lizard_progress/parsers/ribx_parser.py
@@ -138,6 +138,7 @@ def check_gwsw(file_obj):
 class RibxParser(ProgressParser):
     ERRORS = {
         'LOCATION_NOT_FOUND': "Onbekende streng/put/kolk ref '{}'.",
+        'LOCATION_COORD_ERROR': "Incorrect or missing position",
         'X_NOT_IN_EXTENT': "Buiten gebied: X coördinaat niet tussen {} en {}.",
         'Y_NOT_IN_EXTENT': "Buiten gebied: Y coördinaat niet tussen {} en {}.",
         'ATTACHMENT_ALREADY_EXISTS':
@@ -386,7 +387,15 @@ class RibxParser(ProgressParser):
             # y values! Perhaps a better way to do this is is to correctly
             # parse the Points as 2D in the ribxlib...
             point_2d = ogr.Geometry(ogr.wkbPoint)
-            point_2d.AddPoint_2D(item.geom.GetX(), item.geom.GetY())
+
+            # CHANGE 2018-07-04 for PROJ-312
+            # Specify failure reason if location cannot be created because of
+            # wrong or missing coordinates
+            try:
+                point_2d.AddPoint_2D(item.geom.GetX(), item.geom.GetY())
+            except AttributeError:
+                return 'LOCATION_COORD_ERROR'
+
             geom = point_2d
         else:
             geom = item.geom
@@ -482,6 +491,13 @@ class RibxReinigingInspectieRioolParser(RibxParser):
                 return None
             else:
                 location = self.create_new(item)
+
+                # CHANGE 2018-07-04 for PROJ-312
+                if location in self.ERRORS:
+                    self.record_error(item.sourceline,
+                                      location,
+                                      self.ERRORS[location].format(item.ref))
+                return None
 
         manhole_start = None
         if isinstance(item, ribxmodels.InspectionPipe):